\name{dglm.fit}
\alias{dglm.fit}
\title{
  Density function for GLM fit
}
\description{
For internal use within the haplo.stats library
}
\usage{
dglm.fit(fit)
}
\arguments{

\item{fit}{

}
}
\value{
}
\section{Side Effects}{

}
\details{
For internal use within the haplo.stats library
}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
