% $Author: sinnwell $ 
% $Date: 2008/04/29 17:20:59 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/haplo.model.frame.Rd,v 1.2 2008/04/29 17:20:59 sinnwell Exp $ 
% $Locker:  $ 
% $Log: haplo.model.frame.Rd,v $
% Revision 1.2  2008/04/29 17:20:59  sinnwell
% rm allele.lev and miss.val
%
% Revision 1.1  2008/01/09 17:38:24  sinnwell
% Initial revision
%
%Revision 1.3  2004/03/03 22:14:57  schaid
%added allele.lev to work in R with character alleles
%
%Revision 1.2  2004/02/26 22:46:30  sinnwell
%remove example
%
%Revision 1.1  2004/02/26 22:17:14  sinnwell
%Initial revision


\name{haplo.model.frame}
\alias{haplo.model.frame}
\title{
  Sets up a model frame for haplo.glm
}
\description{
For internal use within the haplo.stats library
}
\usage{
haplo.model.frame(m, locus.label=NA, control=haplo.glm.control())
}
\arguments{

\item{m}{
model.frame from evaluated formula
}
\item{locus.label }{
labels for loci in genotype matrix
}
\item{control}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
