% $Author: sinnwell $ 
% $Date: 2008/01/09 19:44:30 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/louis.info.Rd,v 1.1 2008/01/09 19:44:30 sinnwell Exp $ 
% $Locker:  $ 
% $Log: louis.info.Rd,v $
% Revision 1.1  2008/01/09 19:44:30  sinnwell
% Initial revision
%

\name{louis.info}
\alias{louis.info}
\title{
  Louis Information for haplo.glm
}
\description{
For internal use within the haplo.stats library
}
\usage{
louis.info(fit)
}
\arguments{

\item{fit}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
