\name{mf.gindx}
\alias{mf.gindx}
\title{
  Model Frame Genotype Index to Account for Missing Data in haplo.glm
}
\description{
For internal use within the haplo.stats library
}
\usage{
mf.gindx(m)
}
\arguments{

\item{m}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
