\name{print.haplo.glm}
\alias{print.haplo.glm}
\title{
  Print a contents of a haplo.glm object
}
\description{
Print model information and then haplotype information. 
}
\usage{
print.haplo.glm(x, print.all.haplo=FALSE, digits = 
                max(options()$digits - 4, 3), ...)
}
\arguments{

\item{x}{
A haplo.glm object
}
\item{print.all.haplo }{
Logical.  If TRUE, print all haplotypes considered in the model.
}
\item{digits}{
Number of numeric digits to print.
}
\item{\dots }{
Optional arguments for print method
}
}
\value{
Nothing is returned
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
haplo.glm
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
