\name{print.seqhap}
\alias{print.seqhap}
\title{
  Print Contents of a Seqhap Object
}
\description{
Print the results from a sequential haplotype association analysis for
case-control data.
}
\usage{
print.seqhap(x, digits=max(options()$digits-2, 5), ...)
}
\arguments{
\item{x}{
A seqhap object
}
\item{digits}{
Number of significant digits to print for numeric values
}
\item{\dots }{
Additional parameters for the print method
}
}
\value{
Nothing is returned.
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
seqhap
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
