\name{printBanner}
\alias{printBanner}
\title{
  Print a nice banner
}
\description{

}
\usage{
printBanner(str, banner.width=80, char.perline=60, border="=")
}
\arguments{
\item{str}{
character string - a title within the banner
}
\item{banner.width }{
width of banner
}
\item{char.perline }{
number of characters per line for the title
}
\item{border}{
type of character for the border
}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{
printBanner("This is a pretty banner", banner.width=40, char.perline=30)

#========================================
#         This is a pretty banner
#========================================
}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
