\name{residScaledGlmFit}
\alias{residScaledGlmFit}
\title{
  Scaled Residuals for GLM fit
}
\description{
For internal use within the haplo.stats library
}
\usage{
residScaledGlmFit(fit)
}
\arguments{

\item{fit}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
