\name{varfunc.glm.fit}
\alias{varfunc.glm.fit}
\title{
  Variance Function for GLM
}
\description{
For internal use within the haplo.stats library
}
\usage{
varfunc.glm.fit(fit)
}
\arguments{

\item{fit}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
