% $Author: sinnwell $
% $Date: 2008/01/08 22:28:36 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/haplo.em.fitter.Rd,v 1.1 2008/01/08 22:28:36 sinnwell Exp $
% $Locker:  $ 
% $Log: haplo.em.fitter.Rd,v $
% Revision 1.1  2008/01/08 22:28:36  sinnwell
% Initial revision
%
%Revision 1.2  2004/02/26 22:30:08  sinnwell
%remove example
%
%Revision 1.1  2004

\name{haplo.em.fitter}
\alias{haplo.em.fitter}
\title{
  Compute engine for haplotype EM algorithm 
}
\description{
For internal use within the haplo.stats library
}
\usage{
haplo.em.fitter(n.loci, n.subject, weight, geno.vec, n.alleles,
                max.haps, max.iter, loci.insert.order, min.posterior, 
                tol, insert.batch.size, random.start, iseed1, iseed2, 
                iseed3, verbose)
}
\arguments{

\item{n.loci }{

}
\item{n.subject }{

}
\item{weight}{

}
\item{geno.vec }{

}
\item{n.alleles }{

}
\item{max.haps }{

}
\item{max.iter }{

}
\item{loci.insert.order }{

}
\item{min.posterior }{

}
\item{tol}{

}
\item{insert.batch.size }{

}
\item{random.start }{

}
\item{iseed1}{

}
\item{iseed2}{

}
\item{iseed3}{

}
\item{verbose}{

}
}
\value{
}
\section{Side Effects}{

}
\details{
For internal use within the haplo.stats library
}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
