\name{haplo.model.frame}
\alias{haplo.model.frame}
\title{
  Sets up a model frame for haplo.glm
}
\description{
For internal use within the haplo.stats library
}
\usage{
haplo.model.frame(m, locus.label=NA, allele.lev=NULL, miss.val=c(0,NA), 
                  control=haplo.glm.control())
}
\arguments{

\item{m}{

}
\item{locus.label }{

}
\item{allele.lev }{

}
\item{miss.val }{

}
\item{control}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
