%  $Author: sinnwell $
%  $Date: 2008/04/29 15:56:58 $
%  $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/hapPower.demo.Rd,v 1.1 2008/04/29 15:56:58 sinnwell Exp $
%  $Locker:  $

\name{hapPower.demo}
\alias{hapPower.demo}
\title{
  Set of haplotypes and frequencies for power and
  sample size calculations
}
\description{
  An example set of haplotypes and frequencies for power and sample size
  calculations in haplo.power.cc and haplo.power.qt
}
\usage{
 data(hapPower.demo)
}
\format{
\item{ }{
Data Frame with 5 columns for haplotypes and a column for frequencies
}
}
\source{  
}
\section{References}{
 Schaid, DJ. Power and sample size for testing associations of
 haplotypes with complex traits. {\em Ann Hum Genet} (2005) 70:116-130.
}
\keyword{datasets}
% docclass is function
