% $Author: sinnwell $ 
% $Date: 2008/01/09 21:26:12 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/internalHaploStats.Rd,v 1.1 2008/01/09 21:26:12 sinnwell Exp $ 
% $Locker:  $ 
% $Log: internalHaploStats.Rd,v $
% Revision 1.1  2008/01/09 21:26:12  sinnwell
% Initial revision
%

\name{geno.recode}
\alias{allele.recode}
\alias{geno.recode}
\alias{dglm.fit}
\alias{haplo.chistat}
\alias{haplo.enum}
\alias{haplo.score.glm}
\alias{haplo.score.podds}
\alias{mf.gindx}
\alias{residScaledGlmFit}
\alias{sr.class}
\alias{sr.class<-}
\alias{varfunc.glm.fit}
\title{
  Internal functions for the HaploStats package
}
\description{
 Internal function for the HaploStats package
}
\usage{
  
}
\arguments{
}
\value{

}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
