% $Author: sinnwell $ 
% $Date: 2009/09/30 20:05:38 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/internalHaploStats.Rd,v 1.3 2009/09/30 20:05:38 sinnwell Exp $ 
% $Locker:  $ 
% $Log: internalHaploStats.Rd,v $
% Revision 1.3  2009/09/30 20:05:38  sinnwell
% rm keywords, or leave one with one from /doc/KEYWORDS
%
% Revision 1.2  2009/09/30 19:30:06  sinnwell
% remove empty sections
%
% Revision 1.1  2008/01/09 21:26:12  sinnwell
% Initial revision
%

\name{geno.recode}
\alias{allele.recode}
\alias{geno.recode}
\alias{dglm.fit}
\alias{haplo.chistat}
\alias{haplo.enum}
\alias{haplo.score.glm}
\alias{haplo.score.podds}
\alias{mf.gindx}
\alias{residScaledGlmFit}
\alias{sr.class}
\alias{sr.class<-}
\alias{varfunc.glm.fit}
\title{
  Internal functions for the HaploStats package.  See the help file for
  the main functions (haplo.em, haplo.score, haplo.glm) for details on
  some of these functions. 
}
\description{
 Internal function for the HaploStats package
}
% docclass is function
% Converted by Sd2Rd version 37351.
