\name{louis.info}
\alias{louis.info}
\title{
  Louis Information for haplo.glm
}
\description{
For internal use within the haplo.stats library
}
\usage{
louis.info(fit)
}
\arguments{

\item{fit}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
