\name{na.geno.keep}
\alias{na.geno.keep}
\title{
  Find non-missing rows in the genotype matrix of the model.frame
}
\description{
An internal function for the haplo.stats package
}
\usage{
na.geno.keep(m)
}
\arguments{

\item{m}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
