% $Author: sinnwell $ 
% $Date: 2008/01/09 19:46:11 $ 
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/na.geno.keep.Rd,v 1.1 2008/01/09 19:46:11 sinnwell Exp $ 
% $Locker:  $ 
% $Log: na.geno.keep.Rd,v $
% Revision 1.1  2008/01/09 19:46:11  sinnwell
% Initial revision
%


\name{na.geno.keep}
\alias{na.geno.keep}
\title{
  Remove rows with NA in covariates, but keep genotypes with NAs
}
\description{
An internal function for the haplo.stats package
}
\usage{
na.geno.keep(m)
}
\arguments{

\item{m}{

}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
