\name{print.haplo.em}
\alias{print.haplo.em}
\title{
  Print contents of a haplo.em object
}
\description{
Print a data frame with haplotypes and their frequencies.  Also print
likelihood information.
}
\usage{
print.haplo.em(x, ...)
}
\arguments{

\item{x}{
A haplo.em object
}
\item{\dots }{
optional arguments for print
}
}
\value{
Nothing is returned
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
haplo.em
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
