\name{print.haplo.scan}
\alias{print.haplo.scan}
\title{
  Print a haplo.scan object
}
\description{
Print a haplo.scan object
}
\usage{
print.haplo.scan(x, digits=max(options()$digits - 2, 5), ...)
}
\arguments{

\item{x}{
An object created by haplo.scan
}
\item{digits}{
Significant digits shown for numeric data
}
\item{\dots }{
Options parameters for the print function
}
}
\value{
NULL
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
\code{\link{haplo.scan}}
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
