\name{print.haplo.score.slide}
\alias{print.haplo.score.slide}
\title{
  Print the contents of a haplo.score.slide object
}
\description{
Print the data frame returned from haplo.score.slide
}
\usage{
print.haplo.score.slide(x, digits=max(options()$digits - 2, 5), ...)
}
\arguments{

\item{x}{
A haplo.score.slide object
}
\item{digits}{
Number of digits to print for numeric output
}
\item{\dots }{
Optional arguments for the print method
}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{

}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
