% $Author: sinnwell $ 
% $Date: 2011/11/10 15:29:41 $ 
% $Header: /projects/genetics/cvs/cvsroot/haplo.stats/man/print.haplo.score.slide.Rd,v 1.3 2011/11/10 15:29:41 sinnwell Exp $ 
% $Locker:  $ 
% $Log: print.haplo.score.slide.Rd,v $
% Revision 1.3  2011/11/10 15:29:41  sinnwell
% major update to hapglm, minor changes to Rd files, prepare for version 1.5.0 release
%
% Revision 1.2  2008/01/08 21:56:53  sinnwell
% add RCS keywords
%


\name{print.haplo.score.slide}
\alias{print.haplo.score.slide}
\title{
  Print the contents of a haplo.score.slide object
}
\description{
Print the data frame returned from haplo.score.slide
}
\usage{
\method{print}{haplo.score.slide}(x, digits=max(options()$digits - 2, 5), ...)
}
\arguments{

\item{x}{
A haplo.score.slide object
}
\item{digits}{
Number of digits to print for numeric output
}
\item{\dots }{
Optional arguments for the print method
}
}
% docclass is function
% Converted by Sd2Rd version 37351.
