% $Author: sinnwell $ -->
% $Date: 2008/01/08 22:23:19 $ -->
% $Header: /people/biostat3/sinnwell/Haplo/Make/RCS/print.seqhap.Rd,v 1.2 2008/01/08 22:23:19 sinnwell Exp $ -->
% $Locker:  $ -->
% $Log: print.seqhap.Rd,v $
% Revision 1.2  2008/01/08 22:23:19  sinnwell
% add comments
%
%Revision 1.3  2007/04/16 20:11:37  sinnwell
%add ...
%
%Revision 1.2  2007/04/12 21:30:37  sinnwell
%*** empty log message ***
%
%Revision 1.1  2007/04/06 19:31:34  sinnwell
%Initial revision

\name{print.seqhap}
\alias{print.seqhap}
\title{
  Print Contents of a Seqhap Object
}
\description{
Print the results from a sequential haplotype association analysis for
case-control data.
}
\usage{
\method{print}{seqhap}(x, digits=max(options()$digits-2, 5), ...)
}
\arguments{
\item{x}{
A seqhap object
}
\item{digits}{
Number of significant digits to print for numeric values
}
\item{\dots }{
Additional parameters for the print method
}
}
\value{
Nothing is returned.
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
seqhap
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
