\name{summary.haplo.em}
\alias{summary.haplo.em}
\title{
  Summarize contents of a haplo.em object
}
\description{
Display haplotype pairs and their posterior probabilities by subject.
Also display a table with number of max haplotype pairs for a
subject versus how many were kept (max vs. used).  
}
\usage{
summary.haplo.em(object, show.haplo=FALSE, nlines=NULL, ...)
}
\arguments{
\item{object}{
A haplo.em object
}
\item{show.haplo }{
Logical. If TRUE, show the alleles of the haplotype pairs, otherwise show
only the recoded values.
}
\item{nlines}{
To shorten output, print the first 1:nlines rows of the large data frame.  
}
\item{\dots }{
Optional arguments for the summary method
}
}
\value{
}
\section{Side Effects}{

}
\details{

}
\section{References}{

}
\seealso{
\code{\link{haplo.em}}
}
\examples{

}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 37351.
